requirements.txt
streamlit>=1.25.0
torch>=2.0.0
transformers>=4.35.0
pdfplumber>=0.10.3
python-docx>=1.1.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
wordcloud>=1.9.0
pyngrok>=6.0.0
